+++ COT/vm_description/ovf/tests/test_doctests.py
@@ -3,7 +3,7 @@
 # test_doctests.py - test runner for COT doctests
 #
 # July 2016, Glenn F. Matthews
-# Copyright (c) 2016 the COT project developers.
+# Copyright (c) 2016-2017 the COT project developers.
 # See the COPYRIGHT.txt file at the top-level directory of this distribution
 # and at https://github.com/glennmatthews/cot/blob/master/COPYRIGHT.txt.
 #
@@ -14,7 +14,7 @@
 # of COT, including this file, may be copied, modified, propagated, or
 # distributed except according to the terms contained in the LICENSE.txt file.
 
-"""Test runner for COT.ovf doctest tests."""
+"""Test runner for COT.vm_description.ovf doctest tests."""
 
 from doctest import DocTestSuite
 from unittest import TestSuite
@@ -26,6 +26,6 @@ def load_tests(*_):
     For the parameters, see :mod:`unittest`. The parameters are unused here.
     """
     suite = TestSuite()
-    suite.addTests(DocTestSuite('COT.ovf.item'))
-    suite.addTests(DocTestSuite('COT.ovf.utilities'))
+    suite.addTests(DocTestSuite('COT.vm_description.ovf.item'))
+    suite.addTests(DocTestSuite('COT.vm_description.ovf.utilities'))
     return suite