+++ COT/vm_description/ovf/tests/test_hardware.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 #
 # November 2016, Glenn F. Matthews
-# Copyright (c) 2016 the COT project developers.
+# Copyright (c) 2016-2017 the COT project developers.
 # See the COPYRIGHT.txt file at the top-level directory of this distribution
 # and at https://github.com/glennmatthews/cot/blob/master/COPYRIGHT.txt.
 #
@@ -12,10 +12,10 @@
 # of COT, including this file, may be copied, modified, propagated, or
 # distributed except according to the terms contained in the LICENSE.txt file.
 
-"""Unit test cases for COT.ovf.OVFHardware class."""
+"""Unit test cases for COT.vm_description.ovf.OVFHardware class."""
 
 from COT.tests.ut import COT_UT
-from COT.ovf import OVF
+from COT.vm_description.ovf import OVF
 
 
 class TestOVFHardware(COT_UT):